x-gif-maker/requirements.txt
# Core dependencies
Pillow>=10.0.0

# MP4 export support (optional)
imageio>=2.31.0
imageio-ffmpeg>=0.4.9

# Upload functionality
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
